requirements.txt
tensorflow-gpu==1.7
numpy
six
scipy
matplotlib
